requirements.txt
Django==1.7.4
MySQL-python==1.2.3
PIL==1.1.7
Pillow==2.6.1
Unidecode==0.04.17
django-appconf==0.6
django-geoposition==0.2.2
django-guardian==1.2.5
django-imagekit==3.2.4
docutils==0.7
egenix-mx-base==3.1.3
future==0.15.2
netifaces==0.10.4
numpy==1.5.0
pilkit==1.1.12
psycopg2==2.5.4
pycrypto==2.3
pygeoip==0.3.2
python-slugify==0.1.0
reportlab==2.4
six==1.8.0
virtualenv==13.1.2
wsgiref==0.1.2
